+++ server/server.h
@@ -2,7 +2,7 @@
  * @Author: zy 953725892@qq.com
  * @Date: 2022-11-15 23:26:16
  * @LastEditors: zy 953725892@qq.com
- * @LastEditTime: 2022-11-16 01:59:33
+ * @LastEditTime: 2022-11-16 17:45:10
  * @FilePath: /lab3/server/server.h
  * @Description: server相关的数据结构和操作
  * 
@@ -15,17 +15,17 @@
 #include "handle.h"
 #include <pthread.h>
 
-typedef unsigned int uint;
+typedef unsigned short ushort;
 
 typedef struct server {
     char* ip;
-    uint port;
+    ushort port;
     char* dir_root;
     int listen_fd;
 } server;
 
 // 初始化server
-void init_server(server* s, char* ip, uint port, char* dir_root);
+void init_server(server* s, char* ip, ushort port, char* dir_root);
 
 // 对客户端的请求进行处理
 void do_server(server* s);
